Build endpoint to Synchronize a single employee

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Response/YearEnd/DuplicateNamesAndBirthdaysResponse.cs
@@ -1,7 +1,7 @@
 ﻿namespace Demoulas.ProfitSharing.Common.Contracts.Response.YearEnd;
 public class DuplicateNamesAndBirthdaysResponse
 {
-    public required long BadgeNumber { get; set; }
+    public required int BadgeNumber { get; set; }
     public required long SSN { get; set; }
     public string? Name { get; set; }
     public required DateOnly DateOfBirth { get; set; }
