+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Response/YearEnd/PayrollDuplicateSSNResponseDto.cs
@@ -4,7 +4,7 @@ using System.Data.SqlTypes;
 namespace Demoulas.ProfitSharing.Common.Contracts.Response.YearEnd;
 public sealed record PayrollDuplicateSSNResponseDto
 {
-    public required long BadgeNumber { get; set; }
+    public required int BadgeNumber { get; set; }
     public required long SSN { get; set; }
     public string? Name { get; set; }
     public required AddressResponseDto Address { get; set; }
